requirements.txt
amqp==2.6.1
asgiref==3.7.2
attrs==23.1.0
billiard==3.6.4.0
celery==4.4.7
Django==4.2.6
django-debug-toolbar==4.2.0
django-filter==23.3
djangorestframework==3.14.0
flower==0.9.7
gunicorn==21.2.0
humanize==4.8.0
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kombu==4.6.11
packaging==23.2
Pillow==10.1.0
prometheus-client==0.8.0
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
redis==3.5.3
referencing==0.30.2
rpds-py==0.10.6
sqlparse==0.4.4
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
vine==1.3.0
